+++ Bloom.Application/Models/NovaAvaliacaoModel.cs
@@ -8,7 +8,7 @@ namespace Bloom.Application.Models
     public class NovaAvaliacaoModel
     {
         public string Texto { get; set; }
-        public int Classificacao { get; set; }
+        public double Classificacao { get; set; }
         public Guid UsuarioId { get; set; }
         public Guid? FilmeId { get; set; }
         public Guid? SerieId { get; set; }
