+++ Bloom.BLL/Entities/Avaliacao.cs
@@ -9,7 +9,7 @@ namespace Bloom.BLL.Entities
     {
         public Guid Id { get; set; }
         public string Texto { get; set; }
-        public int Classificacao { get; set; }
+        public double Classificacao { get; set; }
         public Guid UsuarioId { get; set; }
         public Guid? FilmeId { get; set; }
         public Guid? SerieId { get; set; }
